Ex01 done but the size of the text box is bad
Maybe it's time to understand how pading, border and EdgeInsets work in
flutter so that I can select the aesthetically pleasing and responsive
sizes to the text and the box around it. Also, there must be a way to
make the font size responsive as well.

modified:   TopBar.dart
modified:   ../pubspec.lock
modified:   ../pubspec.yaml
modified:   ../test/widget_test.dart